# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Fiber API with OAuth2 password flow",
        "title": "Openapi spec",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/tariff/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Retrieves a list of tariffs with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tariffs"
                ],
                "summary": "Retrieves lists tariffs with optional search, type, and pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TariffPaginatedResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tariff/count/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Retrieves a count of tariffs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tariffs"
                ],
                "summary": "Count returns total tariffs count",
                "responses": {
                    "200": {
                        "description": "count: 12345",
                        "schema": {
                            "$ref": "#/definitions/CountResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tariff/create/": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Creates new car",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tariffs"
                ],
                "summary": "Creates a new car",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TariffCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/TariffMessageResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "422": {
                        "description": "detail: Validation errors",
                        "schema": {
                            "$ref": "#/definitions/ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tariff/{id}/detail/": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Retrieve single tariff detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tariffs"
                ],
                "summary": "Retrieve single tariff",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tariff ID (int)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/TariffResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/tariff/{id}/update/": {
            "patch": {
                "security": [
                    {
                        "OAuth2Password": [
                            ""
                        ]
                    }
                ],
                "description": "Updates an existing tariff",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tariffs"
                ],
                "summary": "Update a tariff",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Tariff ID (int)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Request data",
                        "name": "requestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/TariffUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/TariffMessageResponse"
                        }
                    },
                    "400": {
                        "description": "detail: BadRequest - invalid request",
                        "schema": {
                            "$ref": "#/definitions/BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "detail: Unauthorized - Invalid token",
                        "schema": {
                            "$ref": "#/definitions/UnauthorizedResponse"
                        }
                    },
                    "403": {
                        "description": "detail: Permission denied",
                        "schema": {
                            "$ref": "#/definitions/PermissionDeniedResponse"
                        }
                    },
                    "404": {
                        "description": "detail: User not found",
                        "schema": {
                            "$ref": "#/definitions/NotFoundResponse"
                        }
                    },
                    "422": {
                        "description": "detail: Validation errors",
                        "schema": {
                            "$ref": "#/definitions/ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "detail: Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/InternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "BadRequestResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "BadRequest - invalid request"
                }
            }
        },
...
        "CountResponse": {
            "type": "object",
            "required": [
                "count"
            ],
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 12345
                }
            }
        },
...
        "InternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Internal Server Error"
                }
            }
        },
...
        "NotFoundResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Not found"
                }
            }
        },
        "PermissionDeniedResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Permission denied"
                }
            }
        },
...
        "TariffCreateInput": {
            "type": "object",
            "required": [
                "duration",
                "name",
                "price_amount"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "price_amount": {
                    "type": "number"
                }
            }
        },
        "TariffMessageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/TariffResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "TariffPaginatedResponse": {
            "type": "object",
            "required": [
                "limit",
                "page",
                "rows"
            ],
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TariffResponse"
                    }
                }
            }
        },
        "TariffResponse": {
            "type": "object",
            "required": [
                "PriceAmount",
                "createdAt",
                "currency",
                "duration",
                "id",
                "isActive",
                "name"
            ],
            "properties": {
                "PriceAmount": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "isActive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "updatedAt": {
                    "description": "optional, can be nil",
                    "type": "string"
                }
            }
        },
        "TariffUpdateInput": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "price_amount": {
                    "type": "number"
                }
            }
        },
        "UnauthorizedResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Unauthorized - Invalid token"
                }
            }
        },
...
        "ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Validation error"
                },
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "fieldName": "this field is required"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/v1/auth/login",
            "scopes": {
                "read": "Grants read access",
                "write": "Grants write access"
            }
        }
    }
}